refactor(uds-ecu): extract shared UDS app; f103 uses it
Move inline DID table, DTC store, and all handler functions (security_seed,
ecu_reset, ecu_routine, ecu_io, ecu_comm) from f103-uds-ecu/firmware/main.c
into a new board-agnostic examples/common/uds_ecu_app.{c,h}. f103 main.c
now calls uds_ecu_app_fill_config(&cfg, "LABWIRED-F103-UDS") and defines
uds_ecu_app_log() as a thin uart_puts wrapper so the shared security_seed
can print UDS_SEED_SERVED through the board UART. Behavior-preserving: all
three f103 smokes (uds-session-smoke, uds-reset-smoke, uds-smoke) exit 0.
